%Script{}: Add processed source and lines
These fields represent the script source with the contents of the
strings removed.

Why? Because it is easy to mistake the contents of a string for Elixir code,
which may result in incorrectly reporting errors inside them.

I've also changed Rules.TrailingWhitespace to use these processed lines.
This stops us reporting trailing whitespace inside strings as erroneous.

Closes https://github.com/lpil/dogma/issues/12